requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
cryptography==36.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-admin==2.0.1
django-allauth==0.46.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filter==21.1
django-heroku==0.3.1
django-ipware==4.0.0
django-js-asset==1.2.2
django-mptt==0.13.4
django-rest-auth==0.9.5
django-six==1.0.4
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.2
Markdown==3.3.6
oauthlib==3.1.1
Pillow==8.4.0
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
screen==1.0.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
xlwt==1.3.0
zipp==3.6.0
